Refactored the ProvateRoute to its own component
This was made because we dont want uneccassary tasks in App.js
