+++ src/components/NavigationBar.jsx
@@ -32,10 +32,9 @@ export default function NavigationBar () {
       <Nav className='mr-auto ml-auto'>
         <Nav.Link as={NavLink} to='/booking' activeStyle={whiteText} activeClassName='selected-tab'><BsCalendar size='1em' /> Boka </Nav.Link>
         <Nav.Link as={NavLink} to='/profile' activeStyle={whiteText} activeClassName='selected-tab'> <MdAssignment size='1.25em' /> Mina bokningar</Nav.Link>
-        {loggedInRole === 'admin' ?
-          (<Nav.Link as={NavLink} to='/admin' activeStyle={whiteText} activeClassName='selected-tab'> <MdAssignment size='1.25em' /> Administratör sidan</Nav.Link>)
-          : <React.Fragment />
-        }
+        {loggedInRole === 'admin'
+          ? (<Nav.Link as={NavLink} to='/admin' activeStyle={whiteText} activeClassName='selected-tab'> <MdAssignment size='1.25em' /> Administratör sidan</Nav.Link>)
+          : <></>}
       </Nav>
 
       <Nav>
